+++ client/src/components/Home/Home.css
@@ -1,10 +1,11 @@
-ui.container {
-  padding: 1em;
+.example-section-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
-.ui.card {
-  margin: 1em;
-  width: 30%;
-}
-.ui.header {
-  margin-top: 12em;
+
+@media only screen and (max-width: 600px) {
+  .example-section-container {
+    flex-direction: column;
+  }
 }
